Detangle API from concrete implementation
