+++ ui/lib/auth/auth.dart
@@ -1,3 +1,4 @@
 export 'models/auth_provider.dart';
 export 'models/user.dart';
 export 'services/auth_service.dart';
+export 'services/auth_service_factory.dart';
